ast of some forms

+++ src/main/scala/flop/Form.scala
@@ -4,7 +4,7 @@ sealed trait Form
 
 object Form {
 
-  case class NumF(value: Int) extends Form
+  case class NumF(value: Float) extends Form
   case class StrF(value: String) extends Form
   case class SymF(value: String) extends Form
   case class ListF(value: List[Form]) extends Form
